upgrade demo to react-native rn42 and react-native-svg 5.1.5
