+++ demo/android/app/src/main/java/com/victorydemo/MainActivity.java
@@ -1,11 +1,6 @@
-package com.demo;
+package com.victorydemo;
 
 import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends ReactActivity {
 
@@ -15,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "Demo";
+        return "VictoryDemo";
     }
 }
