+++ demo/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Demo</string>
+    <string name="app_name">VictoryDemo</string>
 </resources>
