+++ demo/android/settings.gradle
@@ -1,6 +1,5 @@
-rootProject.name = 'Demo'
+rootProject.name = 'VictoryDemo'
+include ':react-native-svg'
+project(':react-native-svg').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-svg/android')
 
 include ':app'
-
-include ':react-native-svg'
-project(':react-native-svg').projectDir = new File(rootProject.projectDir, '../../node_modules/react-native-svg/android')
