+++ demo/index.android.js
@@ -26,7 +26,7 @@ import {
   VictoryScatter,
   VictoryStack,
   VictoryErrorBar
-} from "../lib";
+} from "victory-native";
 
 const styles = StyleSheet.create({
   container: {
@@ -718,4 +718,4 @@ class Demo extends Component {
   }
 }
 
-AppRegistry.registerComponent("Demo", () => Demo);
+AppRegistry.registerComponent('VictoryDemo', () => Demo);
