+++ demo/ios/VictoryDemo/Images.xcassets/AppIcon.appiconset/Contents.json
